refactor(overlay): move context-chunk merge into ctl, where its siblings live
merge_context_chunk()/merge_context_map() re-derived the same
{"$op": {replace|add|remove|clear}} list-patch algorithm _merge_list_ops()
in ctl/overlay/merge.py already implements for spec.tools and friends --
same op keys, same clear/replace/remove/add order, same "skip if already
present" add semantics.

The actual mistake was where these two functions lived: every real caller
of merge_context_map/merge_context_chunk is in ctl (ctl/overlay/merge.py's
context_merge strategy, and ctl/manifest/mas_agent_merge.py) -- nothing in
runtime calls them. Merging an overlay patch onto a base manifest is an
authoring/composition concern, ctl's domain, the same as
_merge_mapping_ops/_merge_plugin_list_ops next to which they now live.
runtime/boundary/context/manifest_context.py keeps only what's genuinely
needed at agent-execution time: resolve_context_chunk/context_chunks_from_
spec, turning an already-decided value into prompt text regardless of how
it got there.

Moving them into ctl.overlay.merge means they can call the existing
_ops_dict/_merge_list_ops directly -- no new shared module, no
cross-package import to route around. merge_context_chunk is now a 2-line
adapter (coerce a scalar base value to a one-item fragment list, then
delegate) instead of a second copy of the op-application logic.

Relocated the merge_context_chunk/merge_context_map unit tests from
runtime/tests/test_manifest_context.py into runtime/tests/test_overlay_
merge.py (which already tests ctl.overlay.merge, per this repo's existing
convention -- see test_flavour_overlay_merge.py). No behavior change --
full ctl+runtime test suite passes unmodified; the 4 root-level tests/
failures are pre-existing on the unmodified baseline too.
