+++ ctl/src/mas/ctl/overlay/merge.py
@@ -197,8 +197,43 @@ def _merge_list_ops(
                 key = dedupe_key(item)
                 if key not in keys:
                     result.append(item)
-        
+
     return result
+
+
+def _as_fragment_list(value: Any) -> list[Any]:
+    if value is None:
+        return []
+    if isinstance(value, list):
+        return list(value)
+    return [value]
+
+
+def merge_context_chunk(existing: Any, incoming: Any) -> Any:
+    """Merge one spec.context.<key> overlay patch value onto its base value.
+
+    A plain value (string/{ref}/list) fully replaces the base chunk -- implicit
+    replace ergonomics, consistent with every other overlay merge strategy in
+    this repo. A `{"$op": {replace|add|remove|clear}}` value instead operates on
+    the base chunk's fragment list via the same list-patch semantics spec.tools
+    already uses (_merge_list_ops above), coercing a scalar base value to a
+    one-item list first -- so an overlay can add or remove a single fragment
+    (e.g. append one sentence to spec.context.role) without restating the rest.
+    """
+    if _ops_dict(incoming) is None:
+        return deepcopy(incoming)
+    return _merge_list_ops(_as_fragment_list(existing), incoming)
+
+
+def merge_context_map(existing: Any, incoming: dict[str, Any]) -> dict[str, Any]:
+    """Merge an overlay's spec.context patch onto the base spec.context map."""
+    merged = deepcopy(existing) if isinstance(existing, dict) else {}
+    for key, value in incoming.items():
+        if value is None:
+            merged.pop(key, None)
+            continue
+        merged[key] = merge_context_chunk(merged.get(key), value)
+    return merged
 
 
 def _merge_mapping_ops(existing: dict[str, Any], incoming: Any) -> dict[str, Any]:
@@ -277,7 +312,10 @@ def _merge_value_by_meta(existing: Any, incoming: Any, meta: dict[str, Any]) -> 
 
     if strategy == "list_ops":
         existing_list = list(existing or []) if isinstance(existing, list) else []
-        return _merge_list_ops(existing_list, incoming, dedupe_key=None)
+        try:
+            return _merge_list_ops(existing_list, incoming, dedupe_key=None)
+        except ValueError as exc:
+            raise OverlayTargetError(str(exc)) from exc
 
     if strategy == "plugin_list_ops":
         existing_list = list(existing or []) if isinstance(existing, list) else []
@@ -299,8 +337,6 @@ def _merge_value_by_meta(existing: Any, incoming: Any, meta: dict[str, Any]) -> 
 
     if strategy == "context_merge":
         if isinstance(incoming, dict):
-            from mas.runtime.boundary.context.manifest_context import merge_context_map
-
             return merge_context_map(existing, incoming)
         return deepcopy(incoming)
