+++ runtime/src/mas/runtime/boundary/context/manifest_context.py
@@ -5,7 +5,6 @@
 from __future__ import annotations
 
 import logging
-from copy import deepcopy
 from pathlib import Path
 from typing import Any
 
@@ -56,8 +55,9 @@ def resolve_context_chunk(value: Any, *, base_dir: Path) -> str | None:
 
     A list value is a sequence of fragments (each itself a string or {ref}),
     resolved individually and joined with newlines -- lets a chunk be composed
-    from multiple pieces (e.g. a base fragment plus one an overlay appended via
-    merge_context_chunk) instead of one monolithic string.
+    from multiple pieces (e.g. a base fragment plus one an overlay appended,
+    see mas.ctl.overlay.merge.merge_context_chunk) instead of one monolithic
+    string.
     """
     if isinstance(value, str):
         text = value.strip()
@@ -112,67 +112,3 @@ def context_chunks_from_spec(spec: dict[str, Any], *, base_dir: Path) -> list[st
         if text:
             out.append(f"[{key}] {text}")
     return out
-
-
-_CONTEXT_CHUNK_OP_KEYS = {"replace", "add", "remove", "clear"}
-
-
-def _context_chunk_ops(value: Any) -> dict[str, Any] | None:
-    """Recognize the `{"$op": {...}}` (or bare) sugar form on a chunk patch value."""
-    if isinstance(value, dict) and isinstance(value.get("$op"), dict):
-        value = value["$op"]
-    if not isinstance(value, dict):
-        return None
-    if not (set(value) & _CONTEXT_CHUNK_OP_KEYS):
-        return None
-    return value
-
-
-def _as_fragment_list(value: Any) -> list[Any]:
-    if value is None:
-        return []
-    if isinstance(value, list):
-        return list(value)
-    return [value]
-
-
-def merge_context_chunk(existing: Any, incoming: Any) -> Any:
-    """Merge one spec.context.<key> overlay patch value onto its base value.
-
-    A plain value (string/{ref}/list) fully replaces the base chunk -- implicit
-    replace ergonomics, consistent with every other overlay merge strategy in
-    this repo. A `{"$op": {replace|add|remove|clear}}` value instead operates on
-    the base chunk's fragment list (coercing a scalar base value to a one-item
-    list first), so an overlay can add or remove a single fragment -- e.g.
-    append one sentence to spec.context.role -- without restating the rest.
-    """
-    ops = _context_chunk_ops(incoming)
-    if ops is None:
-        return deepcopy(incoming)
-
-    result = [] if ops.get("clear") is True else _as_fragment_list(existing)
-
-    if "replace" in ops:
-        result = list(ops.get("replace") or [])
-
-    if "remove" in ops:
-        to_remove = list(ops.get("remove") or [])
-        result = [item for item in result if item not in to_remove]
-
-    if "add" in ops:
-        for item in list(ops.get("add") or []):
-            if item not in result:
-                result.append(item)
-
-    return result
-
-
-def merge_context_map(existing: Any, incoming: dict[str, Any]) -> dict[str, Any]:
-    """Merge an overlay's spec.context patch onto the base spec.context map."""
-    merged = deepcopy(existing) if isinstance(existing, dict) else {}
-    for key, value in incoming.items():
-        if value is None:
-            merged.pop(key, None)
-            continue
-        merged[key] = merge_context_chunk(merged.get(key), value)
-    return merged
